Add files via upload
This is the first V1 of the engine at this point I have only implemented the board and a Basic Eval Engine. There is no move validation Word generation and dynamic advantage calculator based on the word played. Planning on making it a dynamic eval and game engine with the ability to play and analyze the game with all the features mentioned above and more.